Conflicts:
	src/com/francelabs/i18n/text.properties
	src/com/francelabs/i18n/text_en_US.properties
	src/com/francelabs/i18n/text_fr.properties
	src/com/francelabs/i18n/text_fr_FR.properties

+++ debian7/bin/start-datafari.sh
@@ -33,16 +33,19 @@ if is_running $SOLR_PID_FILE; then
    exit 1
 fi
 
-#if is_running $CASSANDRA_PID_FILE; then
-#   echo "Error : Cassandra seems to be running already with PID $(cat $CASSANDRA_PID_FILE)"
-#   exit 1
-#fi
+if is_running $CASSANDRA_PID_FILE; then
+   echo "Error : Cassandra seems to be running already with PID $(cat $CASSANDRA_PID_FILE)"
+   exit 1
+fi
 
 if  [[ "$STATE" = *installed* ]];
 then
-	echo "Start postgres and add ManifoldCF database"
+	echo "Start postgres and cassandra and add ManifoldCF database"
 	rm -rf "${DATAFARI_HOME}/pgsql/data"
 	mkdir "${DATAFARI_HOME}/pgsql/data"
+	rm -rf "${DATAFARI_HOME}/cassandra/data"
+	mkdir "${DATAFARI_HOME}/cassandra/data"
+	CASSANDRA_INCLUDE=$CASSANDRA_ENV $CASSANDRA_HOME/bin/cassandra -p $CASSANDRA_PID_FILE 1>/dev/null
 	id -u postgres &>/dev/null || useradd postgres
 	chown -R postgres "${DATAFARI_HOME}/pgsql"
 	chmod -R 777 "${DATAFARI_HOME}/logs"
@@ -52,15 +55,17 @@ then
 	su postgres -c "${DATAFARI_HOME}/pgsql/bin/pg_ctl -D ${DATAFARI_HOME}/pgsql/data -l ${DATAFARI_HOME}/logs/pgsql.log start"
 	cd "${DATAFARI_HOME}/mcf/mcf_home"
 	bash "initialize.sh"
+	$CASSANDRA_HOME/bin/cqlsh -f ${DATAFARI_HOME}/bin/common/config/cassandra/tables 
 	sed -i "s/\(STATE *= *\).*/\1initialized/" $INIT_STATE_FILE
 else
 	su postgres -c "${DATAFARI_HOME}/pgsql/bin/pg_ctl -D ${DATAFARI_HOME}/pgsql/data -l ${DATAFARI_HOME}/logs/pgsql.log start"
+	CASSANDRA_INCLUDE=$CASSANDRA_ENV $CASSANDRA_HOME/bin/cassandra -p $CASSANDRA_PID_FILE 1>/dev/null
 fi
 
 
-#CASSANDRA_INCLUDE=$CASSANDRA_ENV $CASSANDRA_HOME/bin/cassandra -p $CASSANDRA_PID_FILE 1>/dev/null
+
 cd $MCF_HOME/../bin
 bash mcf_crawler_agent.sh start
 cd $TOMCAT_HOME/bin
 bash startup.sh
-#SOLR_INCLUDE=$SOLR_ENV $SOLR_INSTALL_DIR/bin/solr start
+SOLR_INCLUDE=$SOLR_ENV $SOLR_INSTALL_DIR/bin/solr start
